Added db context and DB entities

+++ minitwit-backend/Data/Follower.cs
@@ -5,7 +5,7 @@ namespace minitwit_backend.Data;
 
 public partial class Follower
 {
-    public int? WhoId { get; set; }
+    public long? WhoId { get; set; }
 
-    public int? WhomId { get; set; }
+    public long? WhomId { get; set; }
 }
